# senbara-rest/api/openapi/v1/openapi.yaml
openapi: 3.1.1
info:
  title: Senbara REST API v1
  summary: Personal ERP REST API using the Go stdlib, OIDC and PostgreSQL
  description: REST API for a simple personal ERP web application built with the Go standard library, OpenID Connect authentication and PostgreSQL data storage. Designed as a reference for modern REST API development with Go.
  termsOfService: https://felicitas.pojtinger.com/tos/dynamic # Filled out at runtime
  x-privacy-policy: https://felicitas.pojtinger.com/privacy/dynamic # Filled out at runtime
  contact:
    name: Felicitas Pojtinger # Filled out at runtime
    url: https://felicitas.pojtinger.com/imprint # Filled out at runtime
    email: felicitas@pojtinger.com # Filled out at runtime
  license:
    name: AGPL-3.0
    url: https://github.com/pojntfx/senbara/blob/main/LICENSE
  version: 0.1.0
servers:
  - url: http://localhost:1337/ # Filled out at runtime
    description: Local development server # Filled out at runtime
...
paths:
...
  /contacts/{id}:
    get:
      tags:
        - contacts
      summary: Get contact including debts and activities
      operationId: getContact
      security:
        - oidc: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Contact retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactData"
        "403":
          description: Unauthorized
          content:
            text/plain:
              schema:
                type: string
        "500":
          description: An internal error occurred
          content:
            text/plain:
              schema:
                type: string
    delete:
      tags:
        - contacts
      summary: Delete a contact
      operationId: deleteContact
      security:
        - oidc: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Contact deleted successfully
          content:
            application/json:
              schema:
                type: integer
                format: int64
        "403":
          description: Unauthorized
          content:
            text/plain:
              schema:
                type: string
        "500":
          description: An internal error occurred
          content:
            text/plain:
              schema:
                type: string
    put:
      tags:
        - contacts
      summary: Update a contact
      operationId: updateContact
      security:
        - oidc: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                first_name:
                  type: string
                last_name:
                  type: string
                email:
                  type: string
                  format: email
                pronouns:
                  type: string
                nickname:
                  type: string
                birthday:
                  type: string
                  format: date
                address:
                  type: string
                notes:
                  type: string
              required:
                - first_name
                - last_name
                - email
                - pronouns
      responses:
        "200":
          description: Contact updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "403":
          description: Unauthorized
          content:
            text/plain:
              schema:
                type: string
        "500":
          description: An internal error occurred
          content:
            text/plain:
              schema:
                type: string
...
  /activities:
    post:
      tags:
        - activities
      summary: Create a new activity
      operationId: createActivity
      security:
        - oidc: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                contact_id:
                  type: integer
                  format: int64
                name:
                  type: string
                date:
                  type: string
                  format: date
                description:
                  type: string
              required:
                - contact_id
                - name
                - date
      responses:
        "200":
          description: Activity created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
        "403":
          description: Unauthorized
          content:
            text/plain:
              schema:
                type: string
        "500":
          description: An internal error occurred
          content:
            text/plain:
              schema:
                type: string
  /activities/{id}:
    get:
      tags:
        - activities
      summary: Get a specific activity
      operationId: getActivity
      security:
        - oidc: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Activity retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityWithContact"
        "403":
          description: Unauthorized
          content:
            text/plain:
              schema:
                type: string
        "500":
          description: An internal error occurred
          content:
            text/plain:
              schema:
                type: string
    delete:
      tags:
        - activities
      summary: Delete an activity
      operationId: deleteActivity
      security:
        - oidc: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Activity deleted successfully
          content:
            application/json:
              schema:
                type: integer
                format: int64
        "403":
          description: Unauthorized
          content:
            text/plain:
              schema:
                type: string
        "500":
          description: An internal error occurred
          content:
            text/plain:
              schema:
                type: string
    put:
      tags:
        - activities
      summary: Update an activity
      operationId: updateActivity
      security:
        - oidc: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                date:
                  type: string
                  format: date
                description:
                  type: string
              required:
                - name
                - date
      responses:
        "200":
          description: Activity updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
        "403":
          description: Unauthorized
          content:
            text/plain:
              schema:
                type: string
        "500":
          description: An internal error occurred
          content:
            text/plain:
              schema:
                type: string
components:
  schemas:
...
    Contact:
      type: object
      properties:
        id:
          type: integer
          format: int64
        first_name:
          type: string
        last_name:
          type: string
        nickname:
          type: string
        email:
          type: string
          format: email
        pronouns:
          type: string
        birthday:
          type: string
          format: date
          nullable: true
        address:
          type: string
        notes:
          type: string
    ContactData:
      type: object
      properties:
        entry:
          $ref: "#/components/schemas/Contact"
        debts:
          type: array
          items:
            $ref: "#/components/schemas/Debt"
        activities:
          type: array
          items:
            $ref: "#/components/schemas/Activity"
    Debt:
      type: object
      properties:
        id:
          type: integer
          format: int64
        amount:
          type: number
          format: float
        currency:
          type: string
        description:
          type: string
    Activity:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        date:
          type: string
          format: date
        description:
          type: string
    ActivityWithContact:
      type: object
      properties:
        activity_id:
          type: integer
          format: int64
        name:
          type: string
        date:
          type: string
          format: date
        description:
          type: string
        contact_id:
          type: integer
          format: int64
        first_name:
          type: string
        last_name:
          type: string
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: /.well-known/openid-configuration # Filled out at runtime
      x-oidc-dcr-initial-access-token-portal-url: ~ # Filled out at runtime